Added fix_cam_time.py
A python function that subtracts 15 days from the date field - use it cautiously! It can fix the issue in monthly (h0) history files where the date field is the time the data are written and not the center of the averaging period. So in h0s the date may be 2001-02-01 but it refers to the January mean.